Mock_Project_Net03/Mock_Project_Net03.Tests.Unit/Services/ClassServiceTests.cs
using AutoMapper;
using CloudinaryDotNet.Actions;
using FluentAssertions;

using Mock_Project_Net03.Common.Payloads.Responses;
using Mock_Project_Net03.Dtos;
using Mock_Project_Net03.Entities;
using Mock_Project_Net03.Exceptions;
using Mock_Project_Net03.Repositories;
using Mock_Project_Net03.Services;
using NSubstitute;
using NSubstitute.ExceptionExtensions;


namespace Mock_Project_Net03.Tests.Unit.Services
{
    public class ClassServiceTests
    {
        private readonly ClassService _classService;
        private readonly IRepository<Class, int> _classRepository = Substitute.For<IRepository<Class, int>>();
        private readonly IRepository<Attendance, int> _attendanceRepository = Substitute.For<IRepository<Attendance, int>>();
        private readonly IRepository<Class_TrainingUnit, int> _classTrainingUnitRepository = Substitute.For<IRepository<Class_TrainingUnit, int>>();
        private readonly IRepository<User, int> _userRepository = Substitute.For<IRepository<User, int>>();
        private readonly IRepository<Enrollment, int> _enrollmentRepository = Substitute.For<IRepository<Enrollment, int>>();
        private readonly IRepository<TrainingProgram, int> _trainingProgramRepository = Substitute.For<IRepository<TrainingProgram, int>>();
        private readonly IRepository<TrainingProgramUnit, int> _trainingProgramUnitRepository = Substitute.For<IRepository<TrainingProgramUnit, int>>();
        private readonly IRepository<UserRole, int> _roleRepository = Substitute.For<IRepository<UserRole, int>>();
        private readonly IRepository<Semester, int> _semesterRepository = Substitute.For<IRepository<Semester, int>>();
        private readonly IMapper _mapper;
        private readonly IRepository<TrainingProgram_Syllabus, int> _program_syllabus = Substitute.For<IRepository<TrainingProgram_Syllabus, int>>();
        private readonly IRepository<Entities.Syllabus, int> _syllabusRepository = Substitute.For<IRepository<Entities.Syllabus, int>>();

        public ClassServiceTests()
        {
            _mapper = new MapperConfiguration(cfg =>
            {
                cfg.CreateMap<Class, ClassModel>();
                cfg.CreateMap<ClassModel, Class>();
                cfg.CreateMap<Class_TrainingUnit, Class_TrainingUnitModel>();
                cfg.CreateMap<Class_TrainingUnitModel, Class_TrainingUnit>();
                cfg.CreateMap<TrainingProgramUnit, TrainingProgramUnitModel>();
                cfg.CreateMap<TrainingProgramUnitModel, TrainingProgramUnit>();

            }).CreateMapper();
            _classService = new ClassService(
                _classRepository,
                _attendanceRepository,
                _classTrainingUnitRepository,
                _userRepository,
                _enrollmentRepository,
                _trainingProgramRepository,
                _trainingProgramUnitRepository,
                _roleRepository,
                _semesterRepository,
                _syllabusRepository,
                _program_syllabus,
                _mapper
                );
        }

        // Arrange
        static readonly DateTime testDate = DateTime.Now;
        public List<Class> classList = new()
            {
                new() {
                    ClassId = 1,
                    ClassName = "Test1",
                    SemesterId = 1,
                    InstructorId = 1,
                    ProgramId = 1,
                    Status = "Active",
                },
                new() {
                    ClassId = 2,
                    ClassName = "Test2",
                    SemesterId = 1,
                    InstructorId = 1,
                    ProgramId = 1,
                    Status = "Active",
                }
            };

        public List<Semester> semesterList = new()
            {
                new()
                {
                    SemesterID = 1,
                    SemesterName = "SPRING 2024",
                    SemesterStartDate = testDate,
                    SemesterEndDate = testDate
                }
            };

        public List<User> instructorList = new()
            {
                new()
                {
                    UserId = 1,
                    UserName = "TestUser",
                    FullName = "FullName",
                    Gender = "Male",
                    Password = "password",
                    Address = "Test",
                    RoleID = 1,
                    Avatar = "Test",
                    BirthDate = testDate,
                    CreateBy = "Test",
                    CreateDate = testDate,
                    Email = "Test@gmail.com",
                    FSU = "Test",
                    Level = "Test",
                    ModifyBy = "Test",
                    ModifyDate = testDate,
                }
            };

        public List<TrainingProgram> programList = new()
            {
                new()
                {
                    ProgramId = 1,
                    ProgramName = "TestProgram",
                    Version = "1.0",
                    Description = "Test",
                    Status = "Acitve",
                    CreateBy = "Me",
                    CreateDate = testDate,
                    StartDate = testDate,
                    EndDate = testDate,
                    LastModifiedDate = testDate,
                    LastUpdatedBy = "Me",
                }
            };

        public List<TrainingProgramUnit> programUnitList = new()
            {
                new()
                {
                    UnitId = 1,
                    UnitName = "Test",
                    Description = "Test",
                    Index = 1,
                    SyllabusId = 1,
                    Status = "Acitve",
                    Time = 1,
                }
            };

        public List<Attendance> attendees = new()
            {
                new()
                {
                    AttendanceId = 1,
                    ClassId = 1,
                    Date = testDate,
                    Status = "Active",
                    TraineeId = 1
                }
            };
        public List<Class_TrainingUnit> classTrainingUnits = new()
        {
            new()
            {
                Id = 1,
                TrainingProgramUnitId = 1,
                ClassId = 1,
                TrainerId = 1,
                RoomId = 1,
                Slot = 1,
                Day = testDate
            },
        };

        public List<Entities.Syllabus> syllabus = new()
        {
            new()
            {
                SyllabusId = 1,
                TrainingDelivery = "Test"
            }
        };

        [Fact]
        public async Task GetAllClass()
        {
            // Arrange
            var classModelList = new List<ClassModel>();

            _classRepository.GetAll().Returns(classList.AsQueryable());
            _classRepository.CountAsync().Returns(classList.Count);

            var totalItems = await _classRepository.CountAsync();
            //var totalPages = (int)Math.Ceiling((double)totalItems / 10);

            foreach(var item in classList)
            {
                classModelList.Add(_mapper.Map<ClassModel>(item));
            }

            // Act
            var result = _classService.GetAllClass(1, 10);

            // Assert
            // result.Should().BeEquivalentTo((classModelList, totalPages)); seemed like the framework cannot compare and check such complicated object
            result.Should().NotBeNull();
        }

        [Fact]
        public async Task GetAllClassWithFilter_KeyWord()
        {
            var classModelList = new List<ClassModel>();

            var classes = classList
                .Where(x => x.ClassName.ToLower().Contains("test1")).ToList();// class name match keyword
            
            _classRepository.GetAll().Returns(classList.AsQueryable());

            (classModelList, int totalPages )= await _classService.GetAllClass(1, 10);
            
            // Act
            var result = await _classService.GetAllClassWithFilter(classModelList, "test1", "", "", 0);

            // Assert
            result.FirstOrDefault().Should().BeEquivalentTo(_mapper.Map<ClassModel>(classes.FirstOrDefault()));
        }

        [Fact]
        public async Task GetAllClassWithFilter_ReturnNull()
        {
            var classModelList = new List<ClassModel>();

            _classRepository.GetAll().Returns(classList.AsQueryable());

            (classModelList, int totalPages) = await _classService.GetAllClass(1, 10);

            // Act
            Func<Task> action = async() => await _classService.GetAllClassWithFilter(classModelList, "expected to be empty", "", "", 0);

            // Assert
            await action.Should().ThrowAsync<NotFoundException>();
        }

        [Fact]
        public async Task GetClassById()
        {
            var classModelList = new List<ClassModel>();
            
            _classRepository.GetAll().Returns(classList.AsQueryable());
            _classRepository.GetByIdAsync(1).Returns(classList.ElementAtOrDefault(0)); // return the first element
            _semesterRepository.GetByIdAsync(1).Returns(semesterList.ElementAtOrDefault(0));
            _userRepository.GetByIdAsync(1).Returns(instructorList.ElementAtOrDefault(0));
            _trainingProgramRepository.GetByIdAsync(1).Returns(programList.ElementAtOrDefault(0));

            (classModelList, int totalPages) = await _classService.GetAllClass(1, 10);
            var expected = classModelList.ElementAtOrDefault(0);
            expected.Semester = semesterList.ElementAtOrDefault(0);
            expected.Program = programList.ElementAtOrDefault(0);
            expected.Instructor = instructorList.ElementAtOrDefault(0);

            // Act
            var result = await _classService.GetClassById(1);

            // Assert
            result.Should().BeEquivalentTo(expected);
        }

        [Fact]
        public async Task GetClassById_ReturnNull()
        {
            // Act
            Func<Task> action = async() => await _classService.GetClassById(10);

            // Assert
            await action.Should().ThrowAsync<NotFoundException>();
        }

        [Fact]
        public async Task GetClassDetail()
        {
            var classModelList = new List<ClassModel>();

            var response = new GetClassDetailResponse();

            _classRepository.GetAll().Returns(classList.AsQueryable());
            _classRepository.GetByIdAsync(1).Returns(classList.ElementAtOrDefault(0));
            _attendanceRepository.GetAll().Returns(attendees.AsQueryable());
            _semesterRepository.GetByIdAsync(1).Returns(semesterList.ElementAtOrDefault(0));
            _userRepository.GetByIdAsync(1).Returns(instructorList.ElementAtOrDefault(0));
            _trainingProgramRepository.GetByIdAsync(1).Returns(programList.ElementAtOrDefault(0));
            _classTrainingUnitRepository.GetAll().Returns(classTrainingUnits.AsQueryable());
            _syllabusRepository.GetByIdAsync(1).Returns(syllabus.ElementAtOrDefault(0));

            classTrainingUnits.ElementAtOrDefault(0).TrainingProgramUnit = programUnitList.ElementAtOrDefault(0);

            response.Class = _mapper.Map<ClassModel>(classList.ElementAtOrDefault(0));
            response.Class.Semester = semesterList.ElementAtOrDefault(0);
            response.Class.Program = programList.ElementAtOrDefault(0);
            response.Class.Instructor = instructorList.ElementAtOrDefault(0);
            response.ClassDetail = new List<Class_TrainingUnitModel>()
            {
                _mapper.Map<Class_TrainingUnitModel>(classTrainingUnits.ElementAtOrDefault(0))
            };
            response.attendees = attendees.Where(c => c.ClassId == 1).ToList();
            response.AttendeeNumber = response.attendees.Count();
            response.TrainerNumber = 1;
            response.CreatedDate = programList.ElementAt(0).CreateDate == null
                ? DateTime.MinValue
                : (DateTime)programList.ElementAt(0).CreateDate;
            response.ModifiedDate = programList.ElementAt(0).LastModifiedDate == null
                ? DateTime.MinValue
                : (DateTime)programList.ElementAt(0).LastModifiedDate;
            response.CreatedBy = programList.ElementAt(0).CreateBy;
            response.ModifiedBy = programList.ElementAt(0).LastUpdatedBy;
            response.DeliveryType = "Test";
            response.ClassTime = "8:00 - 12:00";

            // Act
            var result = await _classService.GetClassDetail(1);

            // Assert
            result.Should().BeEquivalentTo(response);
        }

        [Fact]
        public async Task AddTrainingProgramToClass()
        {
            _classRepository.GetByIdAsync(1).Returns(classList.ElementAtOrDefault(0));
            _trainingProgramRepository.GetByIdAsync(1).Returns(programList.ElementAtOrDefault(0));
            _userRepository.GetByIdAsync(1).Returns(instructorList.ElementAtOrDefault(0));
            _trainingProgramUnitRepository.GetAll().Returns(programUnitList.AsQueryable());

            // Act
            var result = await _classService.AddTrainingProgramToClass(1, 1, 1, 1, 1);

            // Assert
            result.Should().BeEquivalentTo(classList.ElementAtOrDefault(0));
        }
        [Fact]
        public async Task DeleteClass()
        {
            _classRepository.GetByIdAsync(2).Returns(classList.ElementAtOrDefault(1));
            _attendanceRepository.GetAll().Returns(attendees.AsQueryable());
            _semesterRepository.GetByIdAsync(1).Returns(semesterList.ElementAtOrDefault(0));
            _userRepository.GetByIdAsync(1).Returns(instructorList.ElementAtOrDefault(0));
            _trainingProgramRepository.GetByIdAsync(1).Returns(programList.ElementAtOrDefault(0));
            _classTrainingUnitRepository.GetAll().Returns(classTrainingUnits.AsQueryable());

            // Act
            var result = await _classService.DeleteClass(2);

            // Assert
            result.Should().BeEquivalentTo("Class delete successfully");
        }

        [Fact]
        public async Task DeleteClass_WhenHaveStudentExist()
        {
            _classRepository.GetByIdAsync(1).Returns(classList.ElementAtOrDefault(0));
            _attendanceRepository.GetAll().Returns(attendees.AsQueryable());
            _semesterRepository.GetByIdAsync(1).Returns(semesterList.ElementAtOrDefault(0));
            _userRepository.GetByIdAsync(1).Returns(instructorList.ElementAtOrDefault(0));
            _trainingProgramRepository.GetByIdAsync(1).Returns(programList.ElementAtOrDefault(0));
            _classTrainingUnitRepository.GetAll().Returns(classTrainingUnits.AsQueryable());

            // Act
            Func<Task> action = async() => await _classService.DeleteClass(1);

            // Assert
            await action.Should().ThrowAsync<Exception>("Class still have student studying");
        }

        [Fact]
        public async Task UpdateTrainingCalendar_ReturnsCorrectResult()
        {
            _classTrainingUnitRepository.GetByIdAsync(1).Returns(classTrainingUnits.ElementAtOrDefault(0));

            // Act
            var result = await _classService.UpdateTrainingCalendar(1, false, ClassTime.Morning);

            // Assert
            result.Should().BeEquivalentTo("Training calendar updated successfully");
        }
    }
}
